Modified Syllabus import and added test users
Mofidied the Syllabus import method in SystemModel to prevent updating the course description and overwriting faculty edits. Added a public method that can be run on sites in debug mode (non-production) which will create the four test IDP users in the Syllabus Users Table (T60000001 - T60000004).  T60000001 is also granted administrative permissions.  Finally, created an additional function in System Admin for assigning users to classes as Instructor or Student to help facilitate testing without having to manipulate the DB directly.